Разделит файл со вспомогательными функциями на отдельные
