+++ src/utils/common-utils.js
@@ -1,0 +1,7 @@
+import dayjs from 'dayjs';
+import {
+  DateFormat,
+} from '../const.js';
+
+
+export const getFormatDate = (date, format = DateFormat.HUMANIZE) => date ? dayjs(date).format(format) : '';
